requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pyotp==2.9.0
sqlparse==0.5.0
typing_extensions==4.11.0
pytest-django==4.8.0
pytest==8.3.1
celery==5.4.0
amqp==5.2.0